Move multipeek to its own file
